refactor: generalize modals
