+++ apps/web/components/modals/base/index.tsx
@@ -1,0 +1,3 @@
+import Panel from './panel'
+
+export { Panel }
